Automatic resize of slides using fit.js
When resizing the window of the browser displaying the slides, or when
displaying the sides on a system using a smaller resolution, the slides
were distorted, and did not keep the original picture ratio. This is now
fixed by using the fit.js library for the slides player, now resizing
the slides correctly according to the window size, and placing them in
the center of their container window. Therefore, the HTML template used
by the slides was changed, and the fit.js library is now included as
well with deck.js when the HTML is generated.